Move utils to class
